[MERGE] merged kalypso 11.06 branch into svn head

+++ KalypsoCalculationConnectors/src/org/kalypso/calculation/plc/postprocessing/PLCPostprocessing_Job.java
@@ -88,7 +88,7 @@ public class PLCPostprocessing_Job extends AbstractInternalStatusJob implements 
       final IRasterDataModel riskCalculatedRasterDataModel = (IRasterDataModel) riskCalculatedRasterDataModelWS.getRootFeature().getAdapter( IRasterDataModel.class );
       final IRasterDataModel riskDifferenceRasterDataModel = (IRasterDataModel) riskDifferenceRasterDataModelWS.getRootFeature().getAdapter( IRasterDataModel.class );
 
-      // adding representative (HQ100) inundation coverages to the model
+      // adding representative (highest HQ) inundation coverages to the model
       int highestReturnPeriod = 1;
       for( final IAnnualCoverageCollection coverageCollection : riskStatusQuoRasterDataModel.getWaterlevelCoverageCollection() )
       {
@@ -185,8 +185,9 @@ public class PLCPostprocessing_Job extends AbstractInternalStatusJob implements 
       FileUtils.moveDirectoryToDirectory( riskCalculatedRasterFolderInput, riskFolderCalculated, false );
       FileUtils.moveDirectoryToDirectory( riskCalculatedRasterFolderOutput, riskFolderCalculated, false );
       FileUtils.moveDirectoryToDirectory( riskDifferenceRasterFolderOutput, riskFolderDifference, false );
-      FileUtils.moveDirectoryToDirectory( naResultsFolder, naFolder, false );
 
+      if( naResultsFolder.exists() )
+        FileUtils.moveDirectoryToDirectory( naResultsFolder, naFolder, false );
     }
     catch( final Exception e )
     {
